Remove Maya style tangents from time splines and add tangent conversion functions.
Remove Maya style tangent values from the time spline (and related)
code. Tangents are now only stored in "standard" format which is a
typed slope and a TsTime (aka doule) width. Maya style tangent values
which have a height instead of width, stored values scaled by 3, and
which invert the height of the pre-tangent are no longer used
internally.

To continue to support Maya we have added TsConvertToStandardTangent
and TsConvertFromStandardTangent. These functions take standard slope
and width values and perform optional conversions on them that can
convert between slope and height values, scale by 3, and/or negate the
slope or height value.

Note that Hermite curves do not have an explicit tangent width value.
Instead, the value used is 1/3 the width of the segment to which they
belong. The conversion functions, however, require a width in order to
convert between slope and height values. It is up to the caller to
provide a width value to the conversion routines, even when working
with Hermite curves.

(Internal change: 2343223)
(Internal change: 2343286)