Implement TsSpline::Sample
Sample return piecewise linear approximations of the spline over a
given time span and within a given error tolerance from the
mathematically exact curve.

TsSpline::Sample returns true on success and populates a passed in
TsSplineSamples<Vertex> or TsSplineSamplesWithSources<Vertex> object
where Vertex is one of GfVec2d, GfVec2f, or GfVec2h.  Both of the
TsSplineSamples* tyeps contain a vector of vectors of Vertex. Each
vector of Vertex is a single connected piecewise linear polyline that
approximates some or all of the spline. If a TsSplineSamplesWithSources
was provided, it will also be populated with a vector of
TsSplineSampleSource values, one value per polyline.  The
TsSplineSampleSource indicates which part of the spline generated it.).
A client may choose to display parts of the spline differently based on
whether its source is a loop iteration, extrapolation, or normal knot
interpolation.

The TsTest framework has been extended to work with sampling.
TsTest_TsEvaluator can still call Eval multiple times but it can now
also call Sample and TsTest_Grapher has been updated to draw
TsSplineSamples* objects.

The new test testTsSplineSampling samples each spline in the
TsTest_Museum and samples it 6 ways. First using Sample then using
SampleWithSources. For each of the sampling routines, the full time
span of the spline's knots is sampled, then 4x that span but with a
larger error tolerance, then just the middle 50% of the span but with a
smaller error tolerance.

Finally, the museumGraph script has been extended to allow sampling
with Sample and/or SampleWithSources. These sampled curves can be
overlaid on top of curves sampled with other means (like repeated
Eval() calls).

(Internal change: 2362223)