Implement Hermite curve support
TsSpline curves have a curve type that is either TsCurveBezier or
TsCurveHermite. The curve type determines the interpolation algorithm
that is used for curved segments of the spline.

TsSpline::Sample also works for Hermite curves, but internally it
simply constructs the equivalent Bezier control points and samples
them; the math is equivalent and the Bezier code path is fast and
tested.

(Internal change: 2364451)