Fix a crash when sampling float or half valued splines.
There was a code path that failed to set _times and _knots in
_Sampler::_UnrollInnerLoops that has been fixed and the tests have been
improved to test those code paths.

It turns out that all of our testing was using double valued splines,
so TsTest_TsEvaluator has been taught to produce splines with float and
half values. This highlighted another numeric issue with converting
extreme tangent slopes and code was added to generate reasonable
tangents, even with a reduced value range.

Fixes #3614

(Internal change: 2365974)