Add loop baking to TsSpline.
Implement the methods BakeInnerLoops, GetKnotsWithInnerLoopsBaked, and
GetKnotsWithLoopsBaked.

BakeInnerLoops modifies the spline in place to remove inner loops in
favor of explicit knots. GetKnotsWithInnerLoopsBaked is the same
algorithm but it returns the knots in a TsKnotMap. In all cases,
calling GetKnotsWithInnerLoopsBaked should return the same set of knots
as calling BakeInnerLoops() followed by GetKnots(), but it does not
modify the spline.

GetKnotsWithLoopsBaked is slightly different. If extrapolation loops
are being used, the spline has a virtually infinite number of knots
between time -inf and +inf. It takes a time interval and returns a
TsKnotMap contining the knots that affect that interval. If the ends of
the time interval are between knots, the knot on the far end of the
spline segment containing the end time is included in the result. In
all cases, a spline populated with the results of
GetKnotsWithLoopsBaked() with the same extrapolation settings as the
input spline should evaluate to the same value for all values in the
input time interval (to within reasonable numeric precision limits).

The knots returned from baking extrapolation loops may include knots
that are generated from multiple input knots. Knots at loop boundaries
get their "pre" values from the end of the loop and their "post" values
from the beginning.

There is also a new override for GetKnots() that takes a time interval
argument. This does not bake any loops, it simply returns the subset of
the spline's knots that affect the time interval. This may include the
knot beyond the end of the time interval if the time interval does not
end on a knot.

(Internal change: 2378894)